legacy/python-for-data-engineers-grupo-4-master/pyproject.toml
[tool.poetry]
name = "python-for-data-engineers-grupo-4"
version = "0.1.0"
description = ""
authors = ["Guilherme S. Magalhães <40049979+Guisilcol@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "python_for_data_engineers_grupo_4"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
pandas = "^1.5.1"
bibtexparser = "^1.4.0"
schema = "^0.7.5"
ipykernel = "^6.17.1"
pyyaml = "^6.0"
openpyxl = "^3.0.10"
fastapi = "^0.88.0"
python-multipart = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
